...in progress...

+++ src/xman/exp.py
@@ -124,8 +124,8 @@ class Exp(ExpStruct):
         if self._data.pipeline is None:
             raise NotExistsXManError(f"There's no pipeline in exp `{self}`!")
         if confirm.request(need_confirm, f"ATTENTION! Remove the pipeline of exp `{self}`\n"
-                                          f"(it will also delete all checkpoints and all pipeline"
-                                          f"data)?"):
+                                          f"(it will also delete the pipeline result, "
+                                         f"all checkpoints, and data)?"):
             maker.delete_pipeline(self, self.__pipeline)
             self.__pipeline = None
             self._save()
@@ -146,10 +146,16 @@ class Exp(ExpStruct):
         filesystem.delete_checkpoints_dir(self.location_dir, need_confirm=False)
         return self
 
-    def start(self) -> 'Exp':
+    def start(self, force_after_error: bool = False) -> 'Exp':
         if self.has_manual_result:
             raise IllegalOperationXManError(f"The `{self}` already has a manual result - delete it "
                                             f"with `delete_manual_result()` method first!")
+        pipeline_data = self._data.pipeline
+        if force_after_error:
+            pipeline_data.started = False
+            pipeline_data.error = None
+            pipeline_data.error_stack = None
+            self.update()
         if self.is_ready_for_start:
             if filesystem.has_checkpoints_dir(self.location_dir) and \
                     self.status.status_str == ExpStructStatus.TO_DO:
@@ -157,25 +163,27 @@ class Exp(ExpStruct):
                                                 f"first with `delete_checkpoints()` method!")
             if self.__pipeline is None:
                 self.__pipeline = maker.recreate_pipeline(self)
+            pipeline_data.started = True
+            self._save()
             try:
                 self.__pipeline.start()
             finally:
                 self._save()
                 self.__pipeline._destroy()
                 self.__pipeline = None
-                filesystem.delete_pipeline_run_data(self.location_dir)
-                filesystem.delete_run_timestamp(self.location_dir)
+                if pipeline_data.finished:
+                    filesystem.delete_pipeline_run_data(self.location_dir)
         else:
             self._check_is_not_active()
             if self.is_manual:
                 raise IllegalOperationXManError(f"Can't start the `{self}` as it's manual - use "
                                                 f"`delete_manual_status()` method first!")
-            pipeline_data = self._data.pipeline
             if pipeline_data is None:  # status == 'EMPTY'
                 raise NotExistsXManError(f"`The {self}` doesn't have a pipeline!")
             if pipeline_data.error:  # status == 'ERROR'
                 raise IllegalOperationXManError(
-                    f"The `{self}` has an error during the previous start!")
+                    f"The `{self}` has an error during the previous start! You can use "
+                    f"`force_after_error=True` flag.")
             if pipeline_data.finished:  # status == 'DONE'
                 raise IllegalOperationXManError(f"`The {self}` was already finished!")
         return self
@@ -300,7 +308,7 @@ class Exp(ExpStruct):
                f"{self._data.name} - {self._data.descr}"
 
     def __is_active_by_time_delta(self):
-        run_timestamp = filesystem.load_run_timestamp(self.location_dir)
+        run_timestamp = filesystem.load_pipeline_run_timestamp(self.location_dir)
         if run_timestamp is None:
             return False
         active_buffer = PipelineConfig.active_buffer_colab if platform.is_colab \
